Dedupe `e_clouds` + `e_clock_tower`
