+++ config/splat.pspeu.sttop.yaml
@@ -65,8 +65,8 @@ segments:
       - [0x17738, c, top/e_room_bg]
       - [0x17818, c, top/e_lock_camera]
       - [0x18048, c, top_psp/unk_18048]
-      - [0x18680, c, top_psp/e_clock_tower]
-      - [0x18C10, c, top_psp/e_clouds]
+      - [0x18680, c, top/e_clock_tower]
+      - [0x18C10, c, top/e_clouds]
       - [0x19580, .data, top/e_blue_axe_knight]
       - [0x19760, .data, top/e_stage_name]
       - [0x1A2A0, .data, top/e_red_door]
@@ -89,6 +89,8 @@ segments:
       - [0x3EC70, .data, top/e_room_bg]
       - [0x3ECB0, .data, top/e_lock_camera]
       - [0x3ED30, data]
+      - [0x3ED68, .data, top/e_clock_tower]
+      - [0x3F120, .data, top/e_clouds]
       - [0x3F280, .rodata, top/e_blue_axe_knight]
       - [0x3F2A8, .rodata, top/e_red_door]
       - [0x3F2C0, .rodata, top/e_collect]
@@ -99,4 +101,6 @@ segments:
       - [0x3F4A0, .bss, top/st_update]
       - [0x3F4E0, .bss, top/e_collect]
       - [0x3F500, bss, 3F500]
+      - [0x3FC28, .bss, top/e_clouds]
+      - [0x3FC30, bss, 3FC30]
   - [0x3FC80]
