+++ config/splat.us.strtop.yaml
@@ -63,7 +63,8 @@ segments:
       - [0x1468, .data, e_yorick]
       - [0x151C, .data, e_skull_lord]
       - [0x1530, .data, gen/rooms]
-      - [0x15B4, .data, e_clock_tower]
+      - [0x15B4, .data, e_clouds]
+      - [0x16B4, .data, e_clock_tower]
       - [0x1A64, .data, gen/e_layout]
       - [0x22FC, .data, stage_data]
       - [0xB274, .data, tilemaps]
@@ -97,7 +98,8 @@ segments:
       - [0x2D800, c, e_yorick]
       - [0x2E19C, c, e_skull_lord]
       - [0x2F028, c, prim_helpers]
-      - [0x2F858, c, e_clock_tower]
+      - [0x2F858, c, e_clouds]
+      - [0x2FFBC, c, e_clock_tower]
       - [0x30460, .bss, create_entity]
       - [0x30470, bss]
       - [0x310D8, .bss, 20B48]
