+++ config/splat.us.sttop.yaml
@@ -57,7 +57,8 @@ segments:
       - [0x1964, data]
       - [0x19BC, .data, e_blue_axe_knight]
       - [0x1B78, data]
-      - [0x1BFC, .data, e_clock_tower]
+      - [0x1BFC, .data, e_clouds]
+      - [0x1CFC, .data, e_clock_tower]
       - [0x20AC, data]
       - [0x288E8, .rodata, e_lock_camera]
       - [0x28904, .rodata, e_richter_fight_trigger]
@@ -92,7 +93,8 @@ segments:
       - [0x386B8, c, e_flea_rider]
       - [0x38BB8, c, e_blue_axe_knight]
       - [0x3A3C8, c, prim_helpers]
-      - [0x3ABF8, c, e_clock_tower]
+      - [0x3ABF8, c, e_clouds]
+      - [0x3B354, c, e_clock_tower]
       - [0x3B7F0, bss]
       - [0x3B850, .bss, create_entity]
       - [0x3B860, bss]
