+++ config/symbols.pspeu.sttop.txt
@@ -66,6 +66,7 @@ TOP_EntityUpdates = 0x09256F80;
 g_EInitBreakable = 0x09257038;
 g_EInitObtainable = 0x09257048;
 g_EInitParticle = 0x09257058;
+g_EInitSpawner = 0x09257068;
 g_EInitInteractable = 0x09257078;
 g_EInitUnkId13 = 0x09257088;
 g_EInitLockCamera = 0x09257098;
