+++ src/st/top/e_clouds.c
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "top.h"
+#include "../e_clouds_keep.h"
