--- src/st/top_psp/e_clouds.c
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../top/top.h"
-
-INCLUDE_ASM("st/top_psp/nonmatchings/top_psp/e_clouds", EntityClouds);
